所有接口使用蓝图统一管理，避免app.py文件内容臃肿
